backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
requires-python = ">=3.13"
dependencies = [
    "aiohttp>=3.13.5",
    "fastapi>=0.136.1",
    "langchain>=1.2.15",
    "langchain-community>=0.4.1",
    "langchain-openai>=1.2.0",
    "langgraph>=1.1.9",
    "psycopg2-binary>=2.9.12",
    "pydantic>=2.13.3",
    "python-dotenv>=1.2.2",
    "uvicorn>=0.46.0",
]
